# swagger-specification.json
{
  "openapi": "3.0.0",
  "paths": {
    "/endosos-services/endorsements": {
      "post": {
        "description": "\nFlujo completo de emisión de endoso:\n1. Consulta la póliza al Core (adaptador configurado)\n2. Evalúa las reglas de elegibilidad del tenant\n3. Calcula el costo según la ruta configurada\n4. Persiste en transacción atómica (con auditoría)\n5. Retorna el endoso con su estado final\n\n**Estados posibles de respuesta:**\n- `EMITTED`: Endoso emitido directamente (sin pago ni aprobación)\n- `PENDING_PAYMENT`: Endoso requiere pago previo\n- `PENDING_APPROVAL`: Endoso enviado a flujo de aprobación manual\n    ",
        "operationId": "EndorsementsController_create",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEndorsementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endoso creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "uuid-del-endoso",
                    "policyId": "POL-001",
                    "status": "EMITTED",
                    "endorsementNumber": "END-2025-000001",
                    "calculation": {
                      "proratedAmount": 87.5,
                      "totalCharge": 91,
                      "formula": "(650 - 450) / 365 × 160 días = 87.67"
                    },
                    "emittedAt": "2025-01-15T14:30:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o reglas bloqueantes"
          },
          "404": {
            "description": "Póliza o tenant no encontrado"
          }
        },
        "summary": "Crear y emitir un endoso",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      },
      "get": {
        "description": "Retorna los endosos del tenant activo con filtros y paginación.",
        "operationId": "EndorsementsController_findAll",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "policyId",
            "required": false,
            "in": "query",
            "description": "Filtrar por póliza",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "PENDING_PAYMENT",
                "PENDING_APPROVAL",
                "EMITTED",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "endorsementTypeId",
            "required": false,
            "in": "query",
            "description": "Filtrar por tipo de endoso",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filtrar por texto de búsqueda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de endosos"
          }
        },
        "summary": "Listar endosos del tenant",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      }
    },
    "/endosos-services/endorsements/dashboard/stats": {
      "get": {
        "operationId": "EndorsementsController_getDashboardStats",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estadísticas agregadas"
          }
        },
        "summary": "Obtener estadísticas agregadas para el dashboard",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      }
    },
    "/endosos-services/endorsements/payment-status/{policyId}": {
      "get": {
        "description": "Devuelve el estado actual de la transacción del pago. Ideal para verificar constantemente mediante polling.",
        "operationId": "EndorsementsController_paymentStatus",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "description": "ID o número de la póliza",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar estado del pago para una póliza (Polling)",
        "tags": [
          "Endorsements",
          "Pagos"
        ]
      }
    },
...
    "/endosos-services/endorsements/{id}": {
      "get": {
        "operationId": "EndorsementsController_findOne",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Endoso encontrado"
          },
          "404": {
            "description": "Endoso no encontrado"
          }
        },
        "summary": "Obtener un endoso por ID",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      }
    },
    "/endosos-services/endorsements/{id}/audit": {
      "get": {
        "operationId": "EndorsementsController_getAuditLogs",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traza de auditoría"
          },
          "404": {
            "description": "Endoso no encontrado"
          }
        },
        "summary": "Obtener la traza de auditoría de un endoso",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      }
    },
...
    "/endosos-services/endorsements/calculate": {
      "post": {
        "description": "\nCalcula el costo financiero sin crear el endoso.\nUsado en el Paso 4 del wizard (Cálculo) para mostrar el desglose financiero.\n    ",
        "operationId": "EndorsementsController_calculate",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "UUID del tenant (aseguradora). Ej: a1b2c3d4-...",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateEndorsementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado del cálculo financiero"
          }
        },
        "summary": "Pre-calcular el costo de un endoso",
        "tags": [
          "Endorsements",
          "Endosos"
        ]
      }
    },
...
  },
  "info": {
    "title": "Motor Universal de Endosos — API",
    "description": "## API REST para el Motor Universal de Endosos\n\n### Autenticación Multitenancy\nTodos los endpoints requieren el header **X-Tenant-Id** con el UUID de la aseguradora.\n\n### Arquitectura\nEste motor implementa arquitectura hexagonal (Ports & Adapters):\n- **IPolicyPort**: Adaptador para consultar pólizas del Core del asegurador\n- **ICatalogPort**: Adaptador para catálogos de coberturas\n- La configuración específica por aseguradora vive en JSONB (`tenant_configs.schema`)\n\n### Endpoints disponibles\n- `POST /api/endorsements` — Crear y emitir endoso (transaccional)\n- `GET /api/endorsements` — Listar endosos con filtros\n- `POST /api/endorsements/evaluate` — Pre-evaluar reglas\n- `POST /api/endorsements/calculate` — Pre-calcular costo\n- `GET /api/policies/:id` — Consultar póliza al Core\n- `GET /api/tenants/:id/config` — Obtener configuración del tenant\n- `PUT /api/tenants/:id/config` — Actualizar configuración del tenant",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-Tenant-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id",
        "description": "UUID del tenant (aseguradora)"
      }
    },
    "schemas": {
...
      "CalculateEndorsementDto": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "description": "ID de la póliza",
            "example": "POL-001"
          },
          "routeId": {
            "type": "string",
            "description": "ID de la ruta de endoso",
            "example": "route-basic-to-plata"
          },
          "effectiveDate": {
            "type": "string",
            "description": "Fecha efectiva del endoso",
            "example": "2025-01-15"
          }
        },
        "required": [
          "policyId",
          "routeId",
          "effectiveDate"
        ]
      },
...
    }
  }
}